Fixed a few bugs

+++ tensortools/torch/metrics.py
@@ -5,7 +5,7 @@ def pearson(x: torch.Tensor, y: torch.Tensor, squared: bool = False, inversed: b
     x, y = torch.abs(x), torch.abs(y)
     if squared:
             x, y = torch.square(x), torch.square(y)
-    s = torch.sum((x - torch.mean(x)) * (y - torch.mean(y)) / torch.tensor(x.numel()))
+    s = torch.sum((x - torch.mean(x)) * (y - torch.mean(y)) / torch.tensor(x.numel()).type(x.dtype))
     p = s / (torch.std(x) * torch.std(y))
     return 1 - p if inversed else p
 
@@ -25,7 +25,7 @@ def mse(x: torch.Tensor, y: torch.Tensor) -> torch.float:
 def dot_product(x: torch.Tensor, y: torch.Tensor, normalized: bool = True) -> torch.float:
     """Return the scalar product between the two complex arrays."""
     prod = torch.sum(x * torch.conj(y))
-    norm = torch.sum(torch.abs(x) * torch.abs(y))
+    norm = torch.sum(torch.abs(x) * torch.abs(y)).type(prod.dtype)
     return prod / norm if normalized else prod
